requirements.txt
flask-mysqldb
flask-ext